ensemble gui refactor
