+++ eee/tools/build_ensemble_gui/__init__.py
@@ -1,0 +1,2 @@
+
+from .ensemble_creator import EnsembleCreator
